+++ Cargo.toml
@@ -55,6 +55,7 @@ snafu = { version = "0.8.5", features = ["futures"] }
 tracing = { version = "0.1", features = ["attributes"] }
 tracing-attributes = { version = "0.1.28" }
 embucket_history = { version = "0.1.0", path = "crates/history" }
+validator = { version = "0.20.0", features = ["derive"] }
 
 datafusion = { version = "47.0.0" }
 datafusion-common = { version = "47.0.0" }
