+++ crates/runtime/Cargo.toml
@@ -100,7 +100,7 @@ tower-http = { version = "0.6.1", features = [
 tower-sessions = { version = "0.14.0" }
 tracing = { workspace = true }
 tracing-attributes = { workspace = true }
-url = { version = "2.5.4" }
+url = { workspace = true }
 
 utoipa = { workspace = true }
 utoipa-axum = { workspace = true }
