--- crates/runtime/src/http/metastore/mod.rs
@@ -1,3 +1,0 @@
-pub mod error;
-pub mod handlers;
-pub mod router;
